refactor(rel-import): add index's to carbon components
