+++ src/components/Tag/index.js
@@ -1,0 +1,1 @@
+export { Tag, TagSkeleton } from 'carbon-components-react';
