+++ src/components/TextArea/index.js
@@ -1,0 +1,1 @@
+export { TextArea, TextAreaSkeleton } from 'carbon-components-react';
