+++ src/components/TimePicker/index.js
@@ -1,0 +1,1 @@
+export { TimePicker } from 'carbon-components-react';
