+++ src/components/Toggle/index.js
@@ -1,0 +1,1 @@
+export { Toggle, ToggleSkeleton } from 'carbon-components-react';
