+++ src/components/ToggleSmall/index.js
@@ -1,0 +1,1 @@
+export { ToggleSmall, ToggleSmallSkeleton } from 'carbon-components-react';
