+++ src/components/TooltipDefinition/index.js
@@ -1,0 +1,1 @@
+export { TooltipDefinition } from 'carbon-components-react';
