+++ src/components/TooltipIcon/index.js
@@ -1,0 +1,1 @@
+export { TooltipIcon } from 'carbon-components-react';
