+++ src/components/UnorderedList/index.js
@@ -1,0 +1,1 @@
+export { UnorderedList } from 'carbon-components-react';
